First attempt to refactor SearchInput into modules
