+++ scss/manifest.scss
@@ -69,7 +69,6 @@
 // misc
 @import 'misc/_avatar.scss';
 @import 'misc/_context_menu.scss';
-@import 'misc/_search-input.scss';
 @import 'misc/_delta-buttons.scss';
 @import 'misc/_keyboard_hint_box.scss';
