--- scss/misc/_search-input.scss
@@ -1,54 +1,0 @@
-input.search-input {
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-  word-wrap: normal;
-  -webkit-box-flex: 1;
-  -ms-flex: 1 1 auto;
-  flex: 1 1 auto;
-  margin: 0;
-  line-height: inherit;
-  border: 0px;
-  font-size: 18px;
-  background-color: transparent;
-  caret-color: var(--navBarSearchPlaceholder);
-  color: var(--globalText);
-
-  &.wrap {
-    margin-left: 40px;
-    padding: 10px 0px;
-    width: calc(100% - 80px);
-    .bp4-navbar & {
-      margin-left: 0px;
-    }
-  }
-}
-
-input#chat-list-search {
-  position: absolute;
-  width: calc(30vw - 75px - 10px);
-  left: 69px;
-  color: var(--navBarText);
-  background-color: var(--searchInputBackgroundColor);
-  border-radius: 3px;
-  text-indent: 0px;
-  font-size: 18px;
-  font-weight: normal;
-  margin: 0;
-  margin-right: 0px;
-  margin-left: 0px;
-
-  &::placeholder {
-    user-select: none;
-  }
-}
-
-.bp4-navbar-group > .clear-button {
-  position: absolute;
-  z-index: 10;
-  left: calc(30vw - 30px);
-  &:before {
-    font-weight: normal !important;
-    font-size: medium !important;
-  }
-}
