+++ src/renderer/components/SearchInput/index.ts
@@ -1,0 +1,2 @@
+export { SearchClearButton } from './SearchClearButton'
+export { SearchInput } from './SearchInput'
