+++ src/renderer/components/SearchInput/styles.module.scss
@@ -1,0 +1,68 @@
+.searchInput {
+  -ms-flex: 1 1 auto;
+  -webkit-box-flex: 1;
+  background-color: transparent;
+  border: 0px;
+  caret-color: var(--navBarSearchPlaceholder);
+  color: var(--globalText);
+  flex: 1 1 auto;
+  font-size: 18px;
+  line-height: inherit;
+  margin: 0;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+  word-wrap: normal;
+
+  &.wrap {
+    margin-left: 40px;
+    padding: 10px 0px;
+    width: calc(100% - 80px);
+
+    .bp4-navbar & {
+      margin-left: 0px;
+    }
+  }
+}
+
+.searchClearButton {
+  position: absolute;
+  z-index: 10;
+  left: calc(30vw - 30px);
+
+  &:before {
+    font-weight: normal !important;
+    font-size: medium !important;
+  }
+}
+
+.searchClearButtonHidden {
+  display: none;
+
+  &.bp4-button[class*='bp4-icon-']::before {
+    color: white;
+  }
+
+  &.bp4-button.bp4-minimal:hover {
+    background-color: unset;
+  }
+}
+
+.lalaTest {
+  background-color: var(--searchInputBackgroundColor);
+  border-radius: 3px;
+  color: var(--navBarText);
+  font-size: 18px;
+  font-weight: normal;
+  left: 69px;
+  margin-left: 0px;
+  margin-right: 0px;
+  margin: 0;
+  position: absolute;
+  text-indent: 0px;
+  width: calc(30vw - 75px - 10px);
+
+  &::placeholder {
+    user-select: none;
+  }
+}
